requirements.txt
aiohttp==3.8.6
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
bidict==0.22.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.0
fastapi-socketio==0.0.10
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.10.0
google-crc32c==1.5.0
h11==0.14.0
httptools==0.6.1
idna==3.4
ifaddr==0.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
netifaces==0.10.6
networkx==3.2
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.1.0
pyav==11.2.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyee==11.0.1
pylibsrtp==0.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.2
